+++ apps/io-profile-async/env.example
@@ -17,6 +17,8 @@ BACKEND_INTERNAL_API_KEY=dummy
 FUNCTION_PROFILE_BASE_URL=https://fn=profile:3000
 FUNCTION_PROFILE_API_KEY=dummy
 
+EXPIRED_SESSION_CTA_URL=https://ctaurl
+
 MAIL_FROM=sender@example.com
 MAILHOG_HOSTNAME=localhost
