+++ apps/io-profile-async/package.json
@@ -50,7 +50,7 @@
     "@azure/storage-queue": "^12.16.0",
     "@pagopa/handler-kit": "^1.1.0",
     "@pagopa/handler-kit-azure-func": "^1.2.1",
-    "@pagopa/io-app-email-templates": "^0.1.6",
+    "@pagopa/io-app-email-templates": "^0.3.0",
     "@pagopa/io-functions-commons": "^29.3.0",
     "@pagopa/logger": "^1.0.1",
     "@pagopa/ts-commons": "^13.1.1",
