+++ apps/io-profile-async/src/main.ts
@@ -1,14 +1,12 @@
 import { CosmosClient } from "@azure/cosmos";
 import { getMailerTransporter } from "@pagopa/io-functions-commons/dist/src/mailer";
+import { NonEmptyString } from "@pagopa/ts-commons/lib/strings";
+import { HtmlToTextOptions } from "html-to-text";
 import { getConfigOrThrow } from "./config";
 import { ExpiredSessionAdvisorFunction } from "./functions/expired-session-advisor";
 import { InfoFunction } from "./functions/info";
 import { ExpiredSessionAdvisorQueueMessage } from "./types/expired-session-advisor-queue-message";
 import { buildIoBackendInternalClient } from "./utils/backend-internal-client/dependency";
-import {
-  EXPIRED_SESSION_EMAIL_TITLE,
-  HTML_TO_TEXT_OPTIONS
-} from "./utils/email-utils";
 import { getFetchApi } from "./utils/fetch-utils";
 import { buildFunctionProfileClient } from "./utils/function-profile-client/dependency";
 
@@ -24,6 +22,11 @@ const fetchApi = getFetchApi(config);
 const backendInternalClient = buildIoBackendInternalClient(fetchApi, config);
 const functionProfileClient = buildFunctionProfileClient(fetchApi, config);
 
+const HTML_TO_TEXT_OPTIONS: HtmlToTextOptions = {
+  selectors: [{ selector: "img", format: "skip" }], // Ignore all document images
+  tables: true
+};
+
 export const Info = InfoFunction({
   connectionString: config.AZURE_STORAGE_CONNECTION_STRING,
   db: database
@@ -32,7 +35,8 @@ export const Info = InfoFunction({
 export const ExpiredSessionAdvisor = ExpiredSessionAdvisorFunction({
   from: config.MAIL_FROM,
   htmlToTextOptions: HTML_TO_TEXT_OPTIONS,
-  title: EXPIRED_SESSION_EMAIL_TITLE
+  title: "Importante: la tua sessione é scaduta." as NonEmptyString,
+  ctaUrl: config.EXPIRED_SESSION_CTA_URL
 })({
   backendInternalClient,
   functionProfileClient,
