+++ apps/io-profile-async/src/utils/mailer-transporter/dependency.ts
@@ -1,0 +1,5 @@
+import { getMailerTransporter } from "@pagopa/io-functions-commons/dist/src/mailer";
+
+export type MailerTransporterDependency = {
+  readonly mailerTransporter: ReturnType<typeof getMailerTransporter>;
+};
